requirements.txt
Flask
pandas==1.5.3
numpy==1.23.5
python-chess==0.31.3
gunicorn